Fix(node): use BigInt for coins

+++ src/main/scala/hydrozoa/l1/multisig/tx/deposit/DepositTxBuilder.scala
@@ -14,6 +14,6 @@ trait DepositTxBuilder {
 
 case class DepositTxRecipe(
     deposit: UtxoIdL1, // the only UTXO with deposit funds in user's wallet
-    depositAmount: Int,
+    depositAmount: BigInt,
     datum: DepositDatum // datum for deposit utxo
 )
